move hardware test
